feat: extract init logic to init.js
