+++ packages/design-core/src/main.js
@@ -10,37 +10,6 @@
  *
  */
 
-import { createApp } from 'vue'
-import initSvgs from '@opentiny/tiny-engine-svgs'
-import { setGlobalConfig } from '@opentiny/tiny-engine-controller'
-import i18n from '@opentiny/tiny-engine-controller/js/i18n'
-import App from './App.vue'
-import globalConfig from '../config/lowcode.config'
-import { initMonitor } from '@opentiny/tiny-engine-controller/js/monitor'
-import { injectGlobalComponents } from '@opentiny/tiny-engine-common'
-import { initHttp } from '@opentiny/tiny-engine-http'
-import 'virtual:svg-icons-register'
+import { init } from './init'
 
-import TinyThemeTool from '@opentiny/vue-theme/theme-tool'
-import { tinySmbTheme } from '@opentiny/vue-theme/theme' // SMB 主题
-import { defineEntry } from '@opentiny/tiny-engine-entry'
-import defaultRegistry from '../registry.js'
-
-defineEntry(defaultRegistry)
-initHttp({ env: import.meta.env })
-
-// eslint-disable-next-line no-new
-new TinyThemeTool(tinySmbTheme, 'smbtheme') // 初始化主题
-
-if (import.meta.env.VITE_ERROR_MONITOR === 'true' && import.meta.env.VITE_ERROR_MONITOR_URL) {
-  initMonitor(import.meta.env.VITE_ERROR_MONITOR_URL)
-}
-
-window.TinyGlobalConfig = globalConfig
-setGlobalConfig(globalConfig)
-
-const app = createApp(App)
-
-initSvgs(app)
-window.lowcodeI18n = i18n
-app.use(i18n).use(injectGlobalComponents).mount('#app')
+init()
